requirements.txt
asttokens==2.0.8
backcall==0.2.0
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
keyboard==0.13.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.4
opencv-python==4.8.1.78
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
Pygments==2.13.0
pynput==1.7.6
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
pywin32==306
pyzmq==24.0.1
six==1.16.0
stack-data==0.5.1
tornado==6.2
traitlets==5.4.0
wcwidth==0.2.5
